Implement Koa API
* Add koa package

* More setup

* Add working tests for get and create routes

* Extract test utils to own directory

* Create separate routers
- Rename original router.
- Add indexRouter.
- Update tests.

* Add logger package
- Restructure game files
- Move api and db to src

* Add endpoint tests

* Add remaining tests for games endpoints

* Update controllers to emit error
- Update jest coverage ignore patterns

+++ src/utils/tests/db.js
@@ -14,21 +14,25 @@ const connect = async () => {
 };
 
 const clear = async () => {
-  const collections = mongoose.connection.collections;
-  for (const key in collections) {
-    const collection = collections[key];
-    await collection.deleteMany();
+  if (mongoServer) {
+    const collections = mongoose.connection.collections;
+    for (const key in collections) {
+      const collection = collections[key];
+      await collection.deleteMany();
+    }
   }
 };
 
 const disconnect = async () => {
-  await mongoose.connection.dropDatabase();
-  await mongoose.connection.close();
-  await mongoServer.stop();
+  if (mongoServer) {
+    await mongoose.connection.dropDatabase();
+    await mongoose.connection.close();
+    await mongoServer.stop();
+  }
 };
 
 module.exports = {
-  connect,
   clear,
+  connect,
   disconnect,
 };
